[WIP] update tabinfo from mainpage
