+++ Files/UserControls/MultitaskingControl/HorizontalMultitaskingControl.xaml.cs
@@ -109,7 +109,7 @@ namespace Files.UserControls.MultitaskingControl
             HorizontalTabView.CanReorderTabs = true;
         }
 
-        private void TabStrip_TabStripDrop(object sender, DragEventArgs e)
+        private async void TabStrip_TabStripDrop(object sender, DragEventArgs e)
         {
             HorizontalTabView.CanReorderTabs = true;
             if (!(sender is TabView tabStrip))
@@ -137,7 +137,7 @@ namespace Files.UserControls.MultitaskingControl
 
             var tabViewItemArgs = TabItemArguments.Deserialize(tabViewItemString);
             ApplicationData.Current.LocalSettings.Values[TabDropHandledIdentifier] = true;
-            MainPage.AddNewTabByParam(tabViewItemArgs.InitialPageType, tabViewItemArgs.NavigationArg, index);
+            await MainPage.AddNewTabByParam(tabViewItemArgs.InitialPageType, tabViewItemArgs.NavigationArg, index);
         }
 
         private void TabStrip_TabDragCompleted(TabView sender, TabViewTabDragCompletedEventArgs args)
