+++ Files/UserControls/MultitaskingControl/IMultitaskingControl.cs
@@ -6,17 +6,11 @@ namespace Files.UserControls.MultitaskingControl
 {
     public interface IMultitaskingControl
     {
-        public delegate void CurrentInstanceChangedEventHandler(object sender, CurrentInstanceChangedEventArgs e);
+        public event EventHandler<CurrentInstanceChangedEventArgs> CurrentInstanceChanged;
 
-        public void UpdateSelectedTab(string tabHeader, string workingDirectoryPath, bool isSearchResultsPage);
+        public ObservableCollection<TabItem> Items { get; }
 
-        event CurrentInstanceChangedEventHandler CurrentInstanceChanged;
-
-        ObservableCollection<TabItem> Items { get; }
-
-        List<ITabItem> RecentlyClosedTabs { get; }
-
-        bool RestoredRecentlyClosedTab { get; set; }
+        public List<ITabItem> RecentlyClosedTabs { get; }
 
         public ITabItemContent GetCurrentSelectedTabInstance();
