+++ Files/UserControls/MultitaskingControl/VerticalTabViewControl.xaml.cs
@@ -99,7 +99,7 @@ namespace Files.UserControls.MultitaskingControl
             VerticalTabView.CanReorderTabs = true;
         }
 
-        private void TabStrip_TabStripDrop(object sender, DragEventArgs e)
+        private async void TabStrip_TabStripDrop(object sender, DragEventArgs e)
         {
             VerticalTabView.CanReorderTabs = true;
             if (!(sender is TabView tabStrip))
@@ -127,7 +127,7 @@ namespace Files.UserControls.MultitaskingControl
 
             var tabViewItemArgs = TabItemArguments.Deserialize(tabViewItemString);
             ApplicationData.Current.LocalSettings.Values[TabDropHandledIdentifier] = true;
-            MainPage.AddNewTabByParam(tabViewItemArgs.InitialPageType, tabViewItemArgs.NavigationArg, index);
+            await MainPage.AddNewTabByParam(tabViewItemArgs.InitialPageType, tabViewItemArgs.NavigationArg, index);
         }
 
         private void TabStrip_TabDragCompleted(TabView sender, TabViewTabDragCompletedEventArgs args)
